+++ src/epy_studio/__init__.py
@@ -15,7 +15,7 @@ graphs.
 
 from __future__ import annotations
 
-__version__ = "0.8.1"
+__version__ = "0.8.2"
 __author__ = "Ing. Angel Navarro-Mora M.Sc."
 
 __all__ = ["__author__", "__version__"]